Make percent an f32 instead of a u32

+++ src/qobuz_types/mod.rs
@@ -192,7 +192,7 @@ pub enum PlaylistGenre {
         name: String,
         path: Vec<u32>,
         slug: String,
-        percent: u32,
+        percent: f32,
     },
 }
